requirements.txt
asgiref==3.9.1
boto3==1.42.23
botocore==1.42.23
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
cloudinary==1.44.1
crispy-bootstrap5==2025.6
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==3.1.0
Django==5.2.5
django-allauth==65.13.1
django-anymail==14.0
django-auto-logout==0.5.1
django-ckeditor==6.7.3
django-ckeditor-5==0.2.19
django-cloudinary-storage==0.3.0
django-crispy-forms==2.4
django-daisy==2.0.7
django-dotenv==1.4.2
django-esewa==1.1.0
django-extensions==4.1
django-filter==25.2
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-mathfilters==1.0.0
django-resized==1.0.3
django-social-share==2.3.0
django-static-fontawesome==6.5.2.1
django-storages==1.14.6
django-taggit==6.1.0
django-templated-mail==1.1.1
django-widget-tweaks==1.5.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
Faker==37.8.0
gunicorn==23.0.0
idna==3.11
jmespath==1.0.1
MarkupSafe==3.0.3
numpy==2.4.0
oauthlib==3.3.1
packaging==25.0
paypalrestsdk==1.13.3
pillow==11.3.0
propcache==0.4.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-environ==0.4.54
python3-openid==3.2.0
pytz==2025.2
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
rjsmin==1.2.5
s3transfer==0.16.0
shortuuid==1.0.13
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.3
stripe==14.1.0
tablib==3.9.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
ua-parser-builtins==202601
ujson==5.11.0
urllib3==2.6.2
validate_email==1.3
Werkzeug==3.1.3
whitenoise==6.11.0
zope.interface==8.1.1
